Create Full_Model_Bernoulli_Classification_Simplified
This is the model I'm currently working with. The hierarchical elements are restricted to error terms to allow some studies or observations to be treated as outliers, as well as to alpha, which I'm assuming has to do with experimental conditions that are similar for all the experiments in any given study. There is also a restriction on k1_20 and k2_20 so that k2_20 is always less than k1_20, which I'm hoping will ensure the same inequality across the temperature range and prevent label switching. Initial runs indicate better parameter estimation, but perhaps still poor mixing in JAGS. Waiting on latest results to judge how well it classifies monophasic estimates as belonging to either the fast or slow portion of biphasic decay.