Create Biphasic_non_linear_least_squares_regression.R
Function to feed initial estimates of k1, k2, and alpha to the MCMC codes. Convergence for the formula is a bit tricky, so the user is asked to guess y and x coordinates of the changepoint to allow for informed initial estimates of the parameters. The resulting formula, as well as a monophasic regression, are plotted so that the user may assess how well they appear to fit the data.